requirements.txt
asgiref==3.5.2
certifi==2022.9.14
distlib==0.3.6
Django==4.1.1
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-taggit==3.0.0
filelock==3.8.0
Pillow==9.2.0
pipenv==2022.9.21
platformdirs==2.5.2
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.2
virtualenv==20.16.5
virtualenv-clone==0.5.7
